feat: split admin reports into per-report page routes
